+++ frontend/mulighetsrommet-veileder-flate/src/components/filtrering/Filtertags.module.scss
@@ -1,0 +1,8 @@
+.filtertags {
+  display: flex;
+  justify-content: flex-start;
+  gap: 1rem;
+  height: fit-content;
+  flex-wrap: wrap;
+  align-self: end;
+}
