+++ frontend/mulighetsrommet-veileder-flate/src/views/tiltaksgjennomforing-oversikt/ViewTiltakstypeOversikt.less
@@ -6,7 +6,7 @@
     grid-gap: 1rem;
     min-height: 86vh;
 
-    .filtercontainer {
+    .filtertagsOgHistorikk {
       display: flex;
       justify-content: space-between;
       grid-row: 1;
@@ -14,15 +14,6 @@
 
       .filterknapp {
         margin-top: 1rem;
-      }
-
-      .filtertags {
-        display: flex;
-        justify-content: flex-start;
-        gap: 1rem;
-        height: fit-content;
-        flex-wrap: wrap;
-        align-self: end;
       }
 
       .nav-enhet-tag {
